fix(error-handling): break loop when encountering NoMoreInput
