+++ proxy/Cargo.toml
@@ -4,5 +4,3 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-serde_json = "1.0.111"
-serde = { version = "1.0.195", features = ["derive"] }
